pyproject.toml
[project]
name = "space_hulk_game"
version = "0.1.0"
description = "A text-based adventure game set in the Warhammer 40K universe using crewAI multi-agent framework"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
keywords = ["crewai", "warhammer-40k", "text-adventure", "ai-agents", "game"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "mem0ai>=0.1.0",
    "pyyaml>=6.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "types-pyyaml>=6.0.0",  # Type stubs for PyYAML
]

[project.scripts]
space_hulk_game = "space_hulk_game.main:run"
run_crew = "space_hulk_game.main:run"
train = "space_hulk_game.main:train"
replay = "space_hulk_game.main:replay"
test = "space_hulk_game.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.pyright]
include = ["src", "tests"]
exclude = [
    "venv",
    ".venv",
    "**/__pycache__",
    "**/.pytest_cache",
    "**/node_modules",
    ".git",
    "tmp"
]
pythonVersion = "3.10"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
stubPath = "typings"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "build",
    "dist",
]

# Ignore missing imports for specific third-party libraries without type stubs
[[tool.mypy.overrides]]
module = [
    "crewai.*",
    "mem0.*",
    "litellm.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    "build",
    "dist",
    "tmp",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py files

[tool.bandit]
exclude_dirs = [
    ".venv",
    "venv",
    "tests",
    "tmp",
]
skips = [
    "B101",  # assert_used - we use asserts in tests
    "B601",  # paramiko_calls - if we use paramiko
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
exclude = '''
/(
    \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | build
  | dist
  | tmp
)/
'''
